fix: detect docker compose file when turn things off

+++ scripts/turn-things-off.sh
@@ -1,10 +1,12 @@
 echo "${_group}Turning things off ..."
 
-if [[ -n "$MINIMIZE_DOWNTIME" ]]; then
-  # Stop everything but proxy service
-  $dc rm -fsv $($dc config --services | grep -v -E '^(web)$')
-else
-  $dc down -t $STOP_TIMEOUT --rmi local --remove-orphans
+if [[ -f "$DOCKER_COMPOSE_FILE" ]]; then
+  if [[ -n "$MINIMIZE_DOWNTIME" ]]; then
+    # Stop everything but proxy service
+    $dc rm -fsv $($dc config --services | grep -v -E '^(web)$')
+  else
+    $dc down -t $STOP_TIMEOUT --rmi local --remove-orphans
+  fi
 fi
 
 echo "${_endgroup}"
